Remove chain Redis repo. Factor out Chain.Validate to chain.Validate (pkg func) and call in API contorller. Fix bug: chainRepo.Add() was never called. Remove all calls to chainRepo.Set(): not needed with in-memory repo. Move chainRepo.Remove() from reapers to traverser because the latter adds it. Rename Job.Sequence.Id tag sequenceStartId to sequenceId.
