+++ Gopkg.toml
@@ -3,14 +3,6 @@
   version = "~1.0"
 
 [[constraint]]
-  name = "github.com/alicebob/miniredis"
-  version = "~2.1"
-
-[[constraint]]
-  name = "github.com/gomodule/redigo"
-  version = "~2.0"
-
-[[constraint]]
   name = "github.com/go-test/deep"
   version = "~1.0"
