+++ job-runner/chain/chain_test.go
@@ -1,4 +1,4 @@
-// Copyright 2017-2018, Square, Inc.
+// Copyright 2017-2019, Square, Inc.
 
 package chain
 
@@ -63,82 +63,6 @@ func TestNewChain(t *testing.T) {
 	running := c.Running()
 	if running["job6"].N != 5 { // (jobs 1, 2, 4, 5, + 6) -> NumJobsRun = 5
 		t.Errorf("got NumJobsRun = %d, expected %d", running["job6"].N, 5)
-	}
-}
-
-func TestFirstJobMultiple(t *testing.T) {
-	jc := &proto.JobChain{
-		Jobs: testutil.InitJobs(4),
-		AdjacencyList: map[string][]string{
-			"job1": {"job3"},
-			"job2": {"job3"},
-			"job3": {"job4"},
-		},
-	}
-	c := NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	_, err := c.FirstJob()
-	if err == nil {
-		t.Errorf("expected an error, but did not get one")
-	}
-}
-
-func TestFirstJobOne(t *testing.T) {
-	jc := &proto.JobChain{
-		Jobs: testutil.InitJobs(4),
-		AdjacencyList: map[string][]string{
-			"job1": {"job2", "job3"},
-			"job2": {"job4"},
-			"job3": {"job4"},
-		},
-	}
-	c := NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	expectedFirstJob := jc.Jobs["job1"]
-	firstJob, err := c.FirstJob()
-
-	if !reflect.DeepEqual(firstJob, expectedFirstJob) {
-		t.Errorf("firstJob = %v, expected %v", firstJob, expectedFirstJob)
-	}
-	if err != nil {
-		t.Errorf("err = %s, expected nil", err)
-	}
-}
-
-func TestLastJobMultiple(t *testing.T) {
-	jc := &proto.JobChain{
-		Jobs: testutil.InitJobs(3),
-		AdjacencyList: map[string][]string{
-			"job1": {"job2", "job3"},
-		},
-	}
-	c := NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	_, err := c.lastJob()
-	if err == nil {
-		t.Errorf("expected an error, but did not get one")
-	}
-}
-
-func TestLastJobOne(t *testing.T) {
-	jc := &proto.JobChain{
-		Jobs: testutil.InitJobs(4),
-		AdjacencyList: map[string][]string{
-			"job1": {"job2", "job3"},
-			"job2": {"job4"},
-			"job3": {"job4"},
-		},
-	}
-	c := NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	expectedLastJob := jc.Jobs["job4"]
-	lastJob, err := c.lastJob()
-
-	if !reflect.DeepEqual(lastJob, expectedLastJob) {
-		t.Errorf("lastJob = %v, expected %v", lastJob, expectedLastJob)
-	}
-	if err != nil {
-		t.Errorf("err = %s, expected nil", err)
 	}
 }
 
@@ -498,198 +422,6 @@ func TestSetState(t *testing.T) {
 	c.SetState(proto.STATE_RUNNING)
 	if c.State() != proto.STATE_RUNNING {
 		t.Errorf("State = %d, want %d", c.State(), proto.STATE_RUNNING)
-	}
-}
-
-func TestIndegreeCounts(t *testing.T) {
-	jc := &proto.JobChain{
-		Jobs: testutil.InitJobs(9),
-		AdjacencyList: map[string][]string{
-			"job1": {"job2", "job3"},
-			"job2": {"job4", "job5"},
-			"job3": {"job5", "job6"},
-			"job4": {"job6", "job7"},
-			"job5": {"job6"},
-			"job6": {"job8"},
-			"job7": {"job8"},
-		},
-	}
-	c := NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	expectedCounts := map[string]int{
-		"job1": 0,
-		"job2": 1,
-		"job3": 1,
-		"job4": 1,
-		"job5": 2,
-		"job6": 3,
-		"job7": 1,
-		"job8": 2,
-		"job9": 0,
-	}
-	counts := c.indegreeCounts()
-
-	if !reflect.DeepEqual(counts, expectedCounts) {
-		t.Errorf("counts = %v, want %v", counts, expectedCounts)
-	}
-}
-
-func TestOutdegreeCounts(t *testing.T) {
-	jc := &proto.JobChain{
-		Jobs: testutil.InitJobs(7),
-		AdjacencyList: map[string][]string{
-			"job1": {"job2", "job3"},
-			"job2": {"job4", "job5", "job6"},
-			"job3": {"job5", "job6"},
-			"job4": {"job5", "job6"},
-			"job5": {"job6"},
-			"job6": {"job7"},
-		},
-	}
-	c := NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	expectedCounts := map[string]int{
-		"job1": 2,
-		"job2": 3,
-		"job3": 2,
-		"job4": 2,
-		"job5": 1,
-		"job6": 1,
-		"job7": 0,
-	}
-	counts := c.outdegreeCounts()
-
-	if !reflect.DeepEqual(counts, expectedCounts) {
-		t.Errorf("counts = %v, want %v", counts, expectedCounts)
-	}
-}
-
-func TestIsAcyclic(t *testing.T) {
-	// No cycle in the chain.
-	jc := &proto.JobChain{
-		Jobs: testutil.InitJobs(4),
-		AdjacencyList: map[string][]string{
-			"job1": {"job2", "job3"},
-			"job2": {"job4"},
-			"job3": {"job4"},
-		},
-	}
-	c := NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	expectedIsAcyclic := true
-	isAcyclic := c.isAcyclic()
-
-	if isAcyclic != expectedIsAcyclic {
-		t.Errorf("isAcyclic = %t, want %t", isAcyclic, expectedIsAcyclic)
-	}
-
-	// Cycle from end to beginning of the chain (i.e., there is no first job).
-	jc = &proto.JobChain{
-		Jobs: testutil.InitJobs(4),
-		AdjacencyList: map[string][]string{
-			"job1": {"job2", "job3"},
-			"job2": {"job4"},
-			"job3": {"job4"},
-			"job4": {"job1"},
-		},
-	}
-	c = NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	expectedIsAcyclic = false
-	isAcyclic = c.isAcyclic()
-
-	if isAcyclic != expectedIsAcyclic {
-		t.Errorf("isAcyclic = %t, want %t", isAcyclic, expectedIsAcyclic)
-	}
-
-	// Cycle in the middle of the chain.
-	jc = &proto.JobChain{
-		Jobs: testutil.InitJobs(4),
-		AdjacencyList: map[string][]string{
-			"job1": {"job2", "job3"},
-			"job2": {"job4"},
-			"job3": {"job5"},
-			"job4": {"job5"},
-			"job5": {"job2", "job6"},
-		},
-	}
-	c = NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	expectedIsAcyclic = false
-	isAcyclic = c.isAcyclic()
-
-	if isAcyclic != expectedIsAcyclic {
-		t.Errorf("isAcyclic = %t, want %t", isAcyclic, expectedIsAcyclic)
-	}
-
-	// No cycle, but multiple first jobs and last jobs.
-	jc = &proto.JobChain{
-		Jobs: testutil.InitJobs(5),
-		AdjacencyList: map[string][]string{
-			"job1": {"job3"},
-			"job2": {"job3"},
-			"job3": {"job4", "job5"},
-		},
-	}
-	c = NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	expectedIsAcyclic = true
-	isAcyclic = c.isAcyclic()
-
-	if isAcyclic != expectedIsAcyclic {
-		t.Errorf("isAcyclic = %t, want %t", isAcyclic, expectedIsAcyclic)
-	}
-}
-
-func TestValidateAdjacencyList(t *testing.T) {
-	// Invalid 1.
-	jc := &proto.JobChain{
-		Jobs: testutil.InitJobs(2),
-		AdjacencyList: map[string][]string{
-			"job1": {"job2", "job3"},
-		},
-	}
-	c := NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	expectedValid := false
-	valid := c.adjacencyListIsValid()
-
-	if valid != expectedValid {
-		t.Errorf("valid = %t, expected %t", valid, expectedValid)
-	}
-
-	// Invalid 2.
-	jc = &proto.JobChain{
-		Jobs: testutil.InitJobs(2),
-		AdjacencyList: map[string][]string{
-			"job1": {"job2"},
-			"job7": {},
-		},
-	}
-	c = NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	expectedValid = false
-	valid = c.adjacencyListIsValid()
-
-	if valid != expectedValid {
-		t.Errorf("valid = %t, expected %t", valid, expectedValid)
-	}
-
-	// Valid.
-	jc = &proto.JobChain{
-		Jobs: testutil.InitJobs(3),
-		AdjacencyList: map[string][]string{
-			"job1": {"job2"},
-			"job2": {"job3"},
-		},
-	}
-	c = NewChain(jc, make(map[string]uint), make(map[string]uint), make(map[string]uint))
-
-	expectedValid = true
-	valid = c.adjacencyListIsValid()
-
-	if valid != expectedValid {
-		t.Errorf("valid = %t, expected %t", valid, expectedValid)
 	}
 }
